Update file shredder

+++ file_shredder.sh
@@ -4,8 +4,10 @@ echo "Enter the filepath of the directory you want to shred:"
 read path
 
 for file in $path/*; do
+    if [ -f "$file" ]; then
        echo "Shredding $file"
        shred $file;
+    fi
 done
 
 echo "All files in directory have been shredded"
